pyproject.toml
[tool.poetry]
name = "crotch-bot"
version = "0.1.0"
description = ""
authors = ["Kuba Gasiorowski <kgasiorowski123@gmail.com>"]
readme = "README.md"
#packages = [{include = "crotch_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
discord-py = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
